requirements.txt
python-dotenv
streamlit
langchain
langchain-core
langchain-community
